Expand SQLite home paths before validation
