+++ src-tauri/src/pool_manager.rs
@@ -1,4 +1,5 @@
 use crate::models::ConnectionParams;
+use crate::sqlite_database::expand_sqlite_filename;
 use deadpool_postgres::{Hook as PgHook, HookError as PgHookError, Manager as PgPoolManager, Pool as PgPool};
 use once_cell::sync::Lazy;
 use rustls::client::danger::{HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier};
@@ -14,7 +15,6 @@ use rustls_platform_verifier::BuilderVerifierExt;
 use sha2::{Digest, Sha256};
 use sqlx::{sqlite::SqliteConnectOptions, ConnectOptions, Connection, Executor, MySql, Pool, Sqlite};
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::RwLock;
@@ -650,26 +650,6 @@ fn build_sqlite_connectoptions(params: &ConnectionParams) -> SqliteConnectOption
     SqliteConnectOptions::new().filename(expand_sqlite_filename(params.database.primary()))
 }
 
-fn expand_sqlite_filename(value: &str) -> PathBuf {
-    let home = directories::BaseDirs::new().map(|dirs| dirs.home_dir().to_path_buf());
-    expand_sqlite_filename_with_home(value, home.as_deref())
-}
-
-fn expand_sqlite_filename_with_home(value: &str, home: Option<&Path>) -> PathBuf {
-    let home_relative = if value == "~" {
-        Some("")
-    } else {
-        value
-            .strip_prefix("~/")
-            .or_else(|| value.strip_prefix("~\\"))
-    };
-
-    match (home_relative, home) {
-        (Some(relative), Some(home)) => home.join(relative),
-        _ => PathBuf::from(value),
-    }
-}
-
 /// Return the connection's startup script if it is set and not blank.
 /// Whitespace-only scripts are treated as absent so the per-connection
 /// hook is skipped entirely rather than issuing an empty query.
@@ -1133,37 +1113,5 @@ pub async fn close_all_pools() {
         for (_, pool) in pools.drain() {
             pool.close().await;
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn expands_sqlite_home_prefixes() {
-        let home = PathBuf::from("/home/dev");
-
-        assert_eq!(
-            expand_sqlite_filename_with_home("~/db.sqlite", Some(&home)),
-            home.join("db.sqlite")
-        );
-        assert_eq!(
-            expand_sqlite_filename_with_home("~\\db.sqlite", Some(&home)),
-            home.join("db.sqlite")
-        );
-        assert_eq!(expand_sqlite_filename_with_home("~", Some(&home)), home);
-    }
-
-    #[test]
-    fn leaves_non_home_sqlite_paths_unchanged() {
-        assert_eq!(
-            expand_sqlite_filename_with_home("relative/db.sqlite", None),
-            PathBuf::from("relative/db.sqlite")
-        );
-        assert_eq!(
-            expand_sqlite_filename_with_home("~user/db.sqlite", Some(Path::new("/home/dev"))),
-            PathBuf::from("~user/db.sqlite")
-        );
     }
 }
